tests/requirements.txt
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-timeout==2.2.0
httpx==0.26.0
locust==2.20.0
kafka-python==2.0.2
redis==5.0.1
python-jose==3.3.0
pyjwt==2.8.1
structlog==23.3.0
